requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
discord.py==2.4.0
docutils==0.21.2
Flask==3.0.3
frozenlist==1.4.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.gstreamer==0.3.3
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
MarkupSafe==2.1.5
multidict==6.1.0
pygame==2.6.0
Pygments==2.18.0
pypiwin32==223
python-dotenv==1.0.1
pywin32==306
requests==2.32.3
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.11.1
